// src/api/invoices/api.ts
export interface GetAllInvoicesPaginatedResponse {
    /**
     * 
     * @type {Pagination}
     * @memberof GetAllInvoicesPaginatedResponse
     */
    'pagination'?: Pagination;
    /**
     * 
     * @type {Array<InvoiceResponse>}
     * @memberof GetAllInvoicesPaginatedResponse
     */
    'data'?: Array<InvoiceResponse>;
}
/**
 *
 * @export
 * @interface InvoiceResponse
 */
export interface InvoiceResponse {
    [x: string]: string;
    [x: string]: string;
    [x: string]: string;
    /**
     *
     * @type {string}
     * @memberof InvoiceResponse
     */
    'id'?: string;
    /**
     *
     * @type {string}
     * @memberof InvoiceResponse
     */
    'invoiceId'?: string;
    /**
     *
     * @type {string}
     * @memberof InvoiceResponse
     */
    'invoice_number'?: string;
    /**
     *
     * @type {string}
     * @memberof InvoiceResponse
     */
    'due_date'?: string;
    /**
     *
     * @type {number}
     * @memberof InvoiceResponse
     */
    'total_amount'?: number;
    /**
     *
     * @type {string}
     * @memberof InvoiceResponse
     */
    'serviceName'?: string;
    /**
     *
     * @type {number}
     * @memberof InvoiceResponse
     */
    'servicePrice'?: number;
    /**
     *
     * @type {number}
     * @memberof InvoiceResponse
     */
    'clientView'?: number;
    /**
     *
     * @type {string}
     * @memberof InvoiceResponse
     */
    'clientId'?: string;
    /**
     * The client\'s email address to which invoice can be sent.
     * @type {string}
     * @memberof InvoiceResponse
     */
    'clientEmail'?: string;
    /**
     *
     * @type {string}
     * @memberof InvoiceResponse
     */
    'createdAt'?: string;
    /**
     *
     * @type {string}
     * @memberof InvoiceResponse
     */
    'kind'?: InvoiceResponseKindEnum;
    /**
     *
     * @type {boolean}
     * @memberof InvoiceResponse
     */
    'settled'?: boolean;
    /**
     *
     * @type {string}
     * @memberof InvoiceResponse
     */
    'status'?: InvoiceResponseStatusEnum;
    /**
     *
     * @type {string}
     * @memberof InvoiceResponse
     */
    'documentUrl'?: string;
}

export const InvoiceResponseKindEnum = {
    F: 'F',
    E: 'E'
} as const;

export type InvoiceResponseKindEnum = typeof InvoiceResponseKindEnum[keyof typeof InvoiceResponseKindEnum];
export const InvoiceResponseStatusEnum = {
    Active: 'ACTIVE',
    Canceled: 'CANCELED',
    Inactive: 'INACTIVE'
} as const;

export type InvoiceResponseStatusEnum = typeof InvoiceResponseStatusEnum[keyof typeof InvoiceResponseStatusEnum];

/**
 * 
 * @export
 * @interface Pagination
 */
export interface Pagination {
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'totalItems'?: number;
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'pageSize'?: number;
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'currentPage'?: number;
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'totalPages'?: number;
}
// ...
export interface InvoiceApiDeleteInvoiceRequest {
    /**
     * The unique identifier of the invoice to delete.
     * @type {string}
     * @memberof InvoiceApiDeleteInvoice
     */
    readonly invoiceId: string
}